Fixed tests to follow the pattern and to be in one class
